+++ packages/shared/src/node-jobs/index.ts
@@ -4,6 +4,7 @@ export * from "./errors.js";
 export * from "./fs.js";
 export * from "./manage.js";
 export * from "./meta.js";
+export * from "./net.js";
 export * from "./process.js";
 export * from "./registry.js";
 export * from "./steamcmd.js";